[PF-48] Builder integration changes
- decoupling library
