+++ projects/example-app/src/app/app.component.html
@@ -1,7 +1,7 @@
 <mat-toolbar fxLayout="row" class="app-container mat-toolbar mat-toolbar-single-row">
     <mat-toolbar-row fxFlex>
         <mat-icon id="netgrif-icon" svgIcon="netgrif_logo"></mat-icon>
-        <mat-button-toggle-group [value]="_mode" fxLayoutAlign="start end">
+        <mat-button-toggle-group [(ngModel)]="_mode" fxLayoutAlign="start end">
             <mat-button-toggle mat-button [value]="canvasMode.SELECT" (click)="changeCanvasMode(canvasMode.SELECT)">
                 <mat-icon svgIcon="cursor-default-outline"></mat-icon>
             </mat-button-toggle>
