+++ projects/example-app/src/app/petriflow-control-set/services/control-petriflow-canvas.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { ControlPetriflowCanvasService } from './control-petriflow-canvas.service';
+
+describe('CanvasElementService', () => {
+  let service: ControlPetriflowCanvasService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(ControlPetriflowCanvasService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
