--- projects/petriflow-svg/src/lib/domain/canvas-event-type.ts
@@ -1,7 +1,0 @@
-
-export enum CanvasEventType {
-    CREATE = 'create',
-    DELETE = 'delete',
-    UPDATE = 'update',
-    SELECT = 'select'
-}
