--- projects/petriflow-svg/src/lib/domain/canvas-event-wrapper.ts
@@ -1,9 +1,0 @@
-import {CanvasEventType} from "./canvas-event-type";
-import {PetriflowCanvasElement} from "../svg-elements/petriflow-canvas-element";
-
-export interface CanvasEventWrapper {
-
-    element: PetriflowCanvasElement;
-
-    eventType: CanvasEventType;
-}
