--- projects/petriflow-svg/src/lib/domain/public-api.ts
@@ -1,2 +1,0 @@
-export * from './canvas-element-collection';
-export * from './canvas-event-wrapper';
