--- projects/petriflow-svg/src/lib/factories/public-api.ts
@@ -1,1 +1,0 @@
-export * from './petriflow-canvas-factory.service';
