+++ projects/petriflow-svg/src/lib/public-api.ts
@@ -1,5 +1,3 @@
-export * from './domain/public-api';
-export * from './factories/public-api';
 export * from './services/public-api';
 export * from './svg-elements/public-api';
 
@@ -10,3 +8,4 @@ export * from './petriflow-canvas.service';
 export * from './petriflow-canvas.module';
 export * from './petriflow-canvas.component';
 export * from './common';
+export * from './grid-configuration';
